# docs/swagger.yaml
basePath: /api
definitions:
  go-marketplace_internal_common.ProblemDetails:
    properties:
      detail:
        description: Specific explanation
        type: string
      errors:
        additionalProperties:
          type: string
        description: Field-specific issues
        type: object
      instance:
        description: Request URI
        type: string
      status:
        description: HTTP status
        type: integer
      title:
        description: Short summary
        type: string
      type:
        description: Error category URI
        type: string
    type: object
  go-marketplace_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  internal_core_wallet.TransactionResponse:
    properties:
      amount:
        type: number
      balance_after:
        type: number
      created_at:
        type: string
      description:
        type: string
      direction:
        type: string
      id:
        type: string
      pending_balance_after:
        type: number
      reference_id:
        type: string
      status:
        type: string
      type:
        type: string
      wallet_id:
        type: string
    type: object
  internal_core_wallet.WalletResponse:
    properties:
      balance:
        type: number
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      pending_balance:
        type: number
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      wallet_number:
        type: string
    type: object
  internal_core_wallet.WithdrawRequest:
    properties:
      amount:
        type: number
      description:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a robust e-commerce backend API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Marketplace API
  version: "1.0"
paths:
...
  /wallets/:
    get:
      consumes:
      - application/json
      description: Fetches the authenticated user's digital wallet information.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_core_wallet.WalletResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get wallet
      tags:
      - wallets
    post:
      consumes:
      - application/json
      description: Creates a new digital wallet for the authenticated user.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_core_wallet.WalletResponse'
              type: object
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Create wallet
      tags:
      - wallets
  /wallets/history:
    get:
      consumes:
      - application/json
      description: Fetches a paginated list of transactions for the user's wallet.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_core_wallet.TransactionResponse'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get wallet history
      tags:
      - wallets
  /wallets/withdraw:
    post:
      consumes:
      - application/json
      description: Deducts funds from the user's wallet balance.
      parameters:
      - description: Withdrawal Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_core_wallet.WithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Withdraw from wallet
      tags:
      - wallets
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"